Added Login and Register funcitonality
Ability to make a new user and "sign in" as that user with all
checks and error checking. Additional functionality/flag will be needed
for knowing who is logged in and the ability to sign out. Additionally
the data shown will need to be stored on the firebase database rather
than locally stored. This is the first step towards that.